Remove legacy login page and restructure application layout with new private and public layouts. Introduce upload functionality with dedicated upload and analysis components, enhancing user experience for X-ray analysis. Add header component for navigation and improve file handling utilities.
