+++ app/(Private)/upload/layout.tsx
@@ -1,0 +1,11 @@
+export default function UploadLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      {children}
+    </>
+  );
+}
