+++ app/(Public)/layout.tsx
@@ -1,0 +1,11 @@
+export default function PublicLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="min-h-screen bg-white">
+      {children}
+    </div>
+  );
+}
