+++ app/(Public)/login/page.tsx
@@ -1,0 +1,16 @@
+import LoginForm from '@/components/forms/auth/login/login.form';
+
+export default function LoginPage() {
+  return (
+    <div className="min-h-screen bg-white flex items-center justify-center">
+      <div className="max-w-md w-full space-y-8">
+        <div>
+          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
+            Sign in to your account
+          </h2>
+        </div>
+        <LoginForm />
+      </div>
+    </div>
+  );
+}
