--- app/(Root)/page.tsx
@@ -1,5 +1,0 @@
-import LandingPage from '@/components/landing/LandingPage';
-
-export default function Home() {
-  return <LandingPage />;
-}
