--- app/(Root)/upload/page.tsx
@@ -1,5 +1,0 @@
-import UploadPage from '@/components/upload/UploadPage';
-
-export default function Upload() {
-  return <UploadPage />;
-}
